+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGCTCBDGWMAFCMPVD70120Q5"
+        "01KGED4W2YVTGMT6R22KBS65E0"
     )
     r"""The unique identifier of the data source"""