+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGCTCBDEY1XDE4A3RWE3Q990"
+        "01KGED4W2X6HS7GJY48VBCKA8T"
     )
     r"""The unique identifier of the data source"""